separate app structure logic and inject appstructure action placeholder
